+++ tools/test-js-with-node
@@ -120,6 +120,7 @@ EXEMPT_FILES = make_set(
         "web/src/message_util.js",
         "web/src/message_view_header.js",
         "web/src/message_viewport.js",
+        "web/src/messages_overlay_ui.js",
         "web/src/muted_topics_ui.js",
         "web/src/muted_users_ui.js",
         "web/src/narrow.js",
