+++ web/templates/tooltip_templates.hbs
@@ -160,9 +160,11 @@
 </template>
 <template id="restore-scheduled-message-tooltip-template">
     {{t 'Edit and reschedule message' }}
+    {{tooltip_hotkey_hints "Enter"}}
 </template>
 <template id="delete-scheduled-message-tooltip-template">
     {{t 'Delete scheduled message' }}
+    {{tooltip_hotkey_hints "Backspace"}}
 </template>
 <template id="create-new-stream-tooltip-template">
     {{t 'Create new stream' }}
